refactor: move utility functions to server directory for better organization -common/utilにひとまとめにしていたserver用関数類をdb.tsとauth.tsに機能別に置き直す。
