+++ app/components/common/commonRegister.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { useCallback, useState } from "react"
 import type { SelectPlayer, SelectUmpire, SelectCompetition } from "@/app/lib/db/schema"
-import { getPlayerList, getUmpireList, getCompetitionList } from "@/app/components/common/utils"
+import { getPlayerList, getUmpireList, getCompetitionList } from "@/app/components/server/db"
 
 type CommonRegisterProps = {
   type: "player" | "umpire" | "course" | "competition"
