+++ app/config/page.tsx
@@ -1,4 +1,4 @@
-import { getCompetitionList, getUmpireList, getCourseList } from "@/app/components/common/utils"
+import { getCompetitionList, getUmpireList, getCourseList } from "@/app/components/server/db"
 import { SelectCompetition, SelectCourse, SelectUmpire } from "@/app/lib/db/schema"
 import View from "@/app/config/view"
